refactor(adr-061): classify remaining parser modules as extract
ADR-061 continuation. Permanently classifies the eleven flat metadata-parser
modules that Phase 5's *_metadata.py split left unclassified
(elf_metadata.py, pe_metadata.py, macho_metadata.py, dwarf_metadata.py,
dwarf_advanced.py, sycl_metadata.py, symvers_metadata.py, python_api.py,
python_ext.py, numpy_capi.py, build_mode.py) as `extract` in
architecture/modules.yaml, matching the ADR's own already-measured "7 real
frontends -> extract findings" inventory.

Classifying them surfaced exactly those 7 dependency-direction violations:
cli_compare_release.py, cli_datasources.py, cli_dump_helpers.py (x4), and
cli_resolve.py each imported a parser module directly rather than through
the engine. Fixed by routing each through abicheck.workflows.extraction,
which already owns this re-export role for the sibling extraction helpers
(seed_includes_and_fold_compile_context, resolve_debug_info, etc.) per its
own docstring's "one owner per operation" contract.

Also classifies abicheck/buildsource/scan_levels.py as `model` -- confirmed
stdlib-only imports (pure enums and functions over them), the same shape
already used for evidence_depth.py.

Fixed three test monkeypatch targets in test_cli_dump_helpers_coverage.py
that patched the parser module directly (abicheck.python_ext.
detect_python_extension etc.) -- since perform_elf_dump now imports these
through workflows.extraction, the patch never reached the real call site.
Two of the three were silently defeated (asserting the mocked return value
that was never actually returned); the third (a "must not run" guard) was
silently unprotected. Fixed per workflows/extraction.py's own documented
convention: patch where the call actually resolves.

Verification:
- python scripts/check_architecture.py -> 0 errors
- python scripts/check_ai_readiness.py -> 0 errors, same warning count as base
- python scripts/adr_status_sync.py -> clean
- pytest -k "compare_release or datasources or dump_helpers or cli_resolve or extraction" -> 409 passed
- pytest tests/test_cli_dump_helpers_coverage.py tests/test_architecture_check.py -> 85 passed
